Cleanup code structure.
