+++ graylog2-web-interface/src/views/components/searchbar/queryvalidation/QueryValidation.test.tsx
